🧹 simplify code that communicates with mvd
