+++ providers-sdk/v1/upstream/mvd/mvd.go
@@ -3,6 +3,8 @@
 
 package mvd
 
+import "go.mondoo.com/cnquery/v9/providers-sdk/v1/inventory"
+
 //go:generate protoc --proto_path=../:. --go_out=. --go_opt=paths=source_relative --rangerrpc_out=. mvd.proto
 
 // Determine all Cves of all Advisories
@@ -22,4 +24,20 @@ func (r *VulnReport) Cves() []*CVE {
 		cveList = append(cveList, v)
 	}
 	return cveList
+}
+
+// MvdPlatform converts the inventory.Platform to the
+// platform object we use for MVD
+func NewMvdPlatform(pf *inventory.Platform) *Platform {
+	if pf == nil {
+		return nil
+	}
+	return &Platform{
+		Name:    pf.Name,
+		Release: pf.Version,
+		Build:   pf.Build,
+		Arch:    pf.Arch,
+		Title:   pf.Title,
+		Labels:  pf.Labels,
+	}
 }
