--- providers/os/resources/mondoo_eol.go
@@ -1,61 +1,0 @@
-// Copyright (c) Mondoo, Inc.
-// SPDX-License-Identifier: BUSL-1.1
-
-package resources
-
-import (
-	"context"
-	"errors"
-	"time"
-
-	"go.mondoo.com/cnquery/v9/llx"
-	"go.mondoo.com/cnquery/v9/providers-sdk/v1/resources"
-	"go.mondoo.com/cnquery/v9/providers-sdk/v1/upstream/mvd"
-)
-
-func (s *mqlMondooEol) id() (string, error) {
-	return "product:" + s.Product.Data + ":" + s.Version.Data, nil
-}
-
-func (s *mqlMondooEol) date() (*time.Time, error) {
-	name := s.Product.Data
-	version := s.Version.Data
-
-	upstream := s.MqlRuntime.Upstream
-	if upstream == nil || upstream.ApiEndpoint == "" {
-		return nil, resources.MissingUpstreamError{}
-	}
-
-	// get new advisory report
-	// start scanner client
-	scannerClient, err := newAdvisoryScannerHttpClient(upstream.ApiEndpoint, upstream.Plugins, upstream.HttpClient)
-	if err != nil {
-		return nil, err
-	}
-
-	platformEolInfo, err := scannerClient.IsEol(context.Background(), &mvd.Platform{
-		Name:    name,
-		Release: version,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	if platformEolInfo == nil {
-		return nil, errors.New("no platform eol information available")
-	}
-
-	var eolDate *time.Time
-
-	if platformEolInfo.EolDate != "" {
-		parsedEolDate, err := time.Parse(time.RFC3339, platformEolInfo.EolDate)
-		if err != nil {
-			return nil, errors.New("could not parse eol date: " + platformEolInfo.EolDate)
-		}
-		eolDate = &parsedEolDate
-	} else {
-		eolDate = &llx.NeverFutureTime
-	}
-
-	return eolDate, nil
-}
